[gn][fidl] Fix signedness warning in fidl formatter
[22794->1034/23885 ~58] CXX host-x64-linux-clang/obj/system/host/fidl/formatter/fidl-format.main.cpp.o
../system/host/fidl/formatter/main.cpp:78:16: warning: comparison of integers of different signs: 'int' and 'std::__2::vector<std::__2::basic_string<char>, std::__2::allocator<std::__2::basic_string<char> > >::size_type' (aka 'unsigned long') [-Wsign-compare]
    while (pos < args.size() && args[pos] != "--" && args[pos].find("-") == 0) {
           ~~~ ^ ~~~~~~~~~~~
../system/host/fidl/formatter/main.cpp:92:13: warning: comparison of integers of different signs: 'int' and 'std::__2::vector<std::__2::basic_string<char>, std::__2::allocator<std::__2::basic_string<char> > >::size_type' (aka 'unsigned long') [-Wsign-compare]
    if (pos >= args.size()) {
        ~~~ ^  ~~~~~~~~~~~
2 warnings generated.

BLD-361 #comment [gn][fidl] Fix signedness warning in fidl formatter
ZX-3415 #comment [gn][fidl] Fix signedness warning in fidl formatter

Test: CQ

+++ zircon/system/host/fidl/formatter/main.cpp
@@ -73,7 +73,7 @@ int main(int argc, char* argv[]) {
     std::vector<std::string> args(argv, argv + argc);
     bool in_place = false;
     bool remove_ordinals = false;
-    int pos = 1;
+    size_t pos = 1;
     // Process options
     while (pos < args.size() && args[pos] != "--" && args[pos].find("-") == 0) {
         if (args[pos] == "-i" || args[pos] == "--in-place") {
